LPS-54424 Move CMIS classes from portal-service into module
Move everything in package `com.liferay.portal.kernel.repository.cmis`
into the new API module.
