+++ modules/apps/document-library/document-library-cmis-repository-api/src/com/liferay/document/library/repository/cmis/search/CMISFullTextConjunction.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.repository.cmis.search;
+package com.liferay.document.library.repository.cmis.search;
 
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
@@ -20,9 +20,9 @@ import com.liferay.portal.kernel.util.StringPool;
 import java.util.List;
 
 /**
- * @author Mika Koivisto
+ * @author Iván Zaera
  */
-public class CMISContainsExpression extends CMISJunction {
+public class CMISFullTextConjunction extends CMISJunction {
 
 	@Override
 	public String toQueryFragment() {
@@ -34,7 +34,9 @@ public class CMISContainsExpression extends CMISJunction {
 
 		StringBundler sb = new StringBundler(cmisCriterions.size() * 2 + 1);
 
-		sb.append("CONTAINS('");
+		if (cmisCriterions.size() > 1) {
+			sb.append(StringPool.OPEN_PARENTHESIS);
+		}
 
 		for (int i = 0; i < cmisCriterions.size(); i++) {
 			CMISCriterion cmisCriterion = cmisCriterions.get(i);
@@ -46,7 +48,9 @@ public class CMISContainsExpression extends CMISJunction {
 			sb.append(cmisCriterion.toQueryFragment());
 		}
 
-		sb.append("')");
+		if (cmisCriterions.size() > 1) {
+			sb.append(StringPool.CLOSE_PARENTHESIS);
+		}
 
 		return sb.toString();
 	}
